+++ build_runner/lib/src/server/server.dart
@@ -47,8 +47,16 @@ ServeHandler createServeHandler(WatchImpl watch) {
   var rootPackage = watch.packageGraph.root.name;
   var assetGraphHanderCompleter = Completer<AssetGraphHandler>();
   var assetHandlerCompleter = Completer<AssetHandler>();
+
   watch.reader.then((reader) async {
-    assetHandlerCompleter.complete(AssetHandler(reader, rootPackage));
+    final notFoundDefaultsToPath = watch.notFoundDefaultsTo;
+    final notFoundDefaultsTo =
+        notFoundDefaultsToPath != null && notFoundDefaultsToPath.isNotEmpty
+            ? AssetId(rootPackage, notFoundDefaultsToPath)
+            : null;
+
+    assetHandlerCompleter
+        .complete(AssetHandler(reader, rootPackage, notFoundDefaultsTo));
     assetGraphHanderCompleter
         .complete(AssetGraphHandler(reader, rootPackage, watch.assetGraph!));
   }).catchError((_) {}); // These errors are separately handled.
@@ -300,10 +308,11 @@ window.\$dartLoader.forceLoadModule('packages/build_runner/src/server/build_upda
 class AssetHandler {
   final FinalizedReader _reader;
   final String _rootPackage;
+  final AssetId? notFoundDefaultsTo;
 
   final _typeResolver = MimeTypeResolver();
 
-  AssetHandler(this._reader, this._rootPackage);
+  AssetHandler(this._reader, this._rootPackage, this.notFoundDefaultsTo);
 
   Future<shelf.Response> handle(shelf.Request request, {String rootDir = ''}) =>
       (request.url.path.endsWith('/') || request.url.path.isEmpty)
@@ -332,6 +341,18 @@ class AssetHandler {
                 return shelf.Response.notFound(
                     await _findDirectoryList(assetId));
               }
+
+              final notFoundDefaultsTo = this.notFoundDefaultsTo;
+
+              if (notFoundDefaultsTo != null) {
+                final pathSegments = assetId.pathSegments;
+
+                if (!pathSegments.contains('package') &&
+                    !pathSegments.contains('lib')) {
+                  return _handle(request, notFoundDefaultsTo);
+                }
+              }
+
               return shelf.Response.notFound('Not Found');
             default:
               return shelf.Response.notFound('Not Found');